make nodegraph available only inside the crate
